refactor command
