+++ lib/appium_lib/core/command.rb
@@ -1,5 +1,5 @@
 module Appium
-  class Driver
+  class Core
     module Commands
       COMMAND_NO_ARG = {
         # core
@@ -8,21 +8,6 @@ module Appium
         close_app:                  [:post, 'session/:session_id/appium/app/close'.freeze],
         reset:                      [:post, 'session/:session_id/appium/app/reset'.freeze],
         device_locked?:             [:post, 'session/:session_id/appium/device/is_locked'.freeze],
-
-        # Android
-        open_notifications:         [:post, 'session/:session_id/appium/device/open_notifications'.freeze],
-        toggle_airplane_mode:       [:post, 'session/:session_id/appium/device/toggle_airplane_mode'.freeze],
-        current_activity:           [:get,  'session/:session_id/appium/device/current_activity'.freeze],
-        current_package:            [:get,  'session/:session_id/appium/device/current_package'.freeze],
-        get_system_bars:            [:get,  'session/:session_id/appium/device/system_bars'.freeze],
-        get_display_density:        [:get,  'session/:session_id/appium/device/display_density'.freeze],
-        is_keyboard_shown:          [:get,  'session/:session_id/appium/device/is_keyboard_shown'.freeze],
-        get_network_connection:     [:get,  'session/:session_id/network_connection'.freeze],
-        get_performance_data_types: [:post, 'session/:session_id/appium/performanceData/types'.freeze],
-
-        # iOS
-        device_time:                [:get,  'session/:session_id/appium/device/system_time'.freeze],
-        current_context:            [:get,  'session/:session_id/context'.freeze]
       }.freeze
 
       COMMAND = {
@@ -46,18 +31,9 @@ module Appium
         update_settings:            [:post, 'session/:session_id/appium/settings'.freeze],
         touch_actions:              [:post, 'session/:session_id/touch/perform'.freeze],
         multi_touch:                [:post, 'session/:session_id/touch/multi/perform'.freeze],
-
-        # Android
-        start_activity:             [:post, 'session/:session_id/appium/device/start_activity'.freeze],
-        end_coverage:               [:post, 'session/:session_id/appium/app/end_test_coverage'.freeze],
-        set_network_connection:     [:post, 'session/:session_id/network_connection'.freeze],
-        get_performance_data:       [:post, 'session/:session_id/appium/getPerformanceData'.freeze],
-
-        # iOS
-        touch_id:                   [:post, 'session/:session_id/appium/simulator/touch_id'.freeze],
-        toggle_touch_id_enrollment: [:post, 'session/:session_id/appium/simulator/toggle_touch_id_enrollment'.freeze]
       }.merge(COMMAND_NO_ARG).freeze
 
+      # TODO: Be able to remove?
       COMMANDS_EXTEND_OSS = COMMAND.merge(::Selenium::WebDriver::Remote::OSS::Bridge::COMMANDS).freeze
       COMMANDS_EXTEND_W3C = COMMAND.merge(::Selenium::WebDriver::Remote::W3C::Bridge::COMMANDS).freeze
     end
