+++ lib/appium_lib/search_context.rb
@@ -1,0 +1,10 @@
+require_relative 'core/search_context'
+require_relative 'android/search_context'
+require_relative 'ios/search_context'
+module Appium
+  class Driver
+    module SearchContext
+      FINDERS = {}.merge(::Appium::Core::SearchContext)
+    end
+  end
+end
